simplifies example scripts and adds complete metadata
